refactor: extract resize logic from pipeline.rs to resize.rs
Move ~800 lines of resize-related code (dimension calculation,
fast_image_resize dispatch, pixel format handling, fallback logic,
and associated tests) from pipeline.rs to a new engine/resize.rs
module. This reduces pipeline.rs from 2,285 to 1,493 lines (35%),
keeping it focused on operation validation, optimization, and dispatch.

Changes:
- New src/engine/resize.rs with all resize functions and their tests
- pipeline.rs imports resize functions instead of defining them inline
- engine.rs declares pub(crate) mod resize and re-exports public API
- Updated imports in memory.rs and api.rs to use resize module directly
- Internal helpers made pub(crate) for cross-module access
